packages/api-server/requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dotenv
firebase-admin